requirements.txt
openai==1.22.0
python-dotenv==1.0.1
streamlit==1.33.0
